requirements.txt
noise==1.2.2
numpy==1.16.4
Pillow==6.0.0
Pytess==1.0.0
